+++ apps/test-app/src/app/app.config.ts
@@ -2,6 +2,7 @@ import { ApplicationConfig, provideZoneChangeDetection } from "@angular/core";
 import { provideRouter } from "@angular/router";
 import { provideAnimationsAsync } from "@angular/platform-browser/animations/async";
 import { provideSsvCommandOptions } from "@ssv/ngx.command";
+import { provideSsvUxViewportOptions } from "@ssv/ngx.ux";
 
 import { appRoutes } from "./app.routes";
 
@@ -15,5 +16,20 @@ export const appConfig: ApplicationConfig = {
 			executingCssClass: "is-busy",
 			hasDisabledDelay: false
 		}),
+
+		// withViewportSsrDevice(() => DeviceType.mobile),
+		provideSsvUxViewportOptions({
+			// breakpoints: {
+			// 	small: 1000,
+			// }
+		}),
+		// provideSsvUxViewportOptions(defaults => {
+		// 	return {
+		// 		breakpoints: {
+		// 			...defaults.breakpoints,
+		// 			small: 1000,
+		// 		}
+		// 	};
+		// }),
 	],
 };