+++ libs/ngx.ux/src/viewport/index.ts
@@ -1,14 +1,15 @@
 export * from "./viewport-data/index";
+
 export { SsvUxViewportModule } from "./viewport.module";
+export { provideSsvUxViewportOptions, VIEWPORT_OPTIONS, type UxViewportOptions, withViewportSsrDevice } from "./viewport.options";
 
 export { SsvViewportMatcherVarDirective, SsvViewportMatcherVarContext } from "./viewport-matcher-var.directive";
 export { SsvViewportMatcherDirective, SsvViewportMatcherContext } from "./viewport-matcher.directive";
-export { UX_VIEWPORT_SSR_DEVICE, ViewportServerSizeService } from "./viewport-server-size.service";
+export { VIEWPORT_SSR_DEVICE, ViewportServerSizeService } from "./viewport-server-size.service";
 export {
 	type ComparisonOperationValueType,
 	type ComparisonOperationLiteral,
 	type ComparisonOperationKeyType,
-	type UxViewportOptions,
 	type ViewportSize,
 	type ViewportSizeTypeInfo,
 	ComparisonOperation,
@@ -17,4 +18,3 @@ export {
 } from "./viewport.model";
 export { ViewportService } from "./viewport.service";
 export { generateViewportSizeType } from "./viewport.util";
-export { UX_VIEWPORT_DEFAULT_BREAKPOINTS } from "./viewport.const";